# backend/docs/swagger/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "Lightweight proxy core management panel for Xray, Sing-box, and Mihomo.",
        "title": "Isolate Panel API",
        "contact": {
            "name": "Isolate Panel",
            "url": "https://github.com/isolate-project/isolate-panel/issues"
        },
        "license": {
            "name": "MIT",
            "url": "https://opensource.org/licenses/MIT"
        },
        "version": "1.0.0"
    },
    "host": "localhost:8080",
    "basePath": "/api",
    "paths": {
...
        "/outbounds": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Returns all outbounds, optionally filtered by core_id or protocol",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "outbounds"
                ],
                "summary": "List outbounds",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Filter by core ID",
                        "name": "core_id",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Filter by protocol name",
                        "name": "protocol",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            },
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Create a new proxy outbound routing rule",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "outbounds"
                ],
                "summary": "Create outbound",
                "parameters": [
                    {
                        "description": "Outbound configuration",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/github_com_isolate-project_isolate-panel_internal_models.Outbound"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            }
        },
        "/outbounds/{id}": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Returns a single outbound by ID",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "outbounds"
                ],
                "summary": "Get outbound",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Outbound ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            },
            "put": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Update outbound fields (partial update)",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "outbounds"
                ],
                "summary": "Update outbound",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Outbound ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Fields to update",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            },
            "delete": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Delete an outbound routing rule",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "outbounds"
                ],
                "summary": "Delete outbound",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Outbound ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            }
        },
        "/protocols": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Returns all registered protocol schemas (25+ protocols). Filter by core or direction.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "protocols"
                ],
                "summary": "List protocols",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Filter by core (xray, sing-box, mihomo)",
                        "name": "core",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Filter by direction (inbound, outbound, both)",
                        "name": "direction",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            }
        },
        "/protocols/{name}": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Returns the full parameter schema for a specific protocol (field types, validation, etc.)",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "protocols"
                ],
                "summary": "Get protocol schema",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Protocol name (e.g. vless, vmess, trojan)",
                        "name": "name",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            }
        },
        "/protocols/{name}/defaults": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Returns auto-generated default values for all protocol parameters (UUID, port, etc.)",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "protocols"
                ],
                "summary": "Get protocol defaults",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Protocol name",
                        "name": "name",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "github_com_isolate-project_isolate-panel_internal_models.Core": {
            "type": "object",
            "properties": {
                "config_path": {
                    "type": "string"
                },
                "created_at": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "is_enabled": {
                    "type": "boolean"
                },
                "is_running": {
                    "type": "boolean"
                },
                "last_error": {
                    "type": "string"
                },
                "log_path": {
                    "type": "string"
                },
                "name": {
                    "description": "singbox, xray, mihomo",
                    "type": "string"
                },
                "pid": {
                    "type": "integer"
                },
                "restart_count": {
                    "type": "integer"
                },
                "updated_at": {
                    "type": "string"
                },
                "uptime_seconds": {
                    "description": "Statistics",
                    "type": "integer"
                },
                "version": {
                    "type": "string"
                }
            }
        },
...
        "github_com_isolate-project_isolate-panel_internal_models.Outbound": {
            "type": "object",
            "properties": {
                "config_json": {
                    "type": "string"
                },
                "core": {
                    "description": "Relations",
                    "allOf": [
                        {
                            "$ref": "#/definitions/github_com_isolate-project_isolate-panel_internal_models.Core"
                        }
                    ]
                },
                "core_id": {
                    "type": "integer"
                },
                "created_at": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "is_enabled": {
                    "type": "boolean"
                },
                "name": {
                    "type": "string"
                },
                "priority": {
                    "type": "integer"
                },
                "protocol": {
                    "type": "string"
                },
                "updated_at": {
                    "type": "string"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "BearerAuth": {
            "description": "JWT Bearer token. Format: \"Bearer {token}\"",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}